FreeBSD: Install noto rather than sub-packages
The noto-basic package includes Noto Mono, but not its newer replacement
Noto Sans Mono, which is part of noto-extra. Noto Sans Mono has some
major improvements that make it better for programming, such as a
slashed zero.

The noto meta-package also includes sub-packages that add support for
not only Japanese, but also Korean and HK, Traditional, and Simplified
Chinese. Even though I'm unlikely to ever write Korean or Chinese, it's
nice to not have mojibake issues for those languages.